BackEnd/requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
joblib==1.4.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
wheel==0.45.1
